+++ inst/NEWS.md
@@ -1,3 +1,11 @@
+# dynutils 1.0.3 (unreleased)
+
+ * FUNCTIONALITY: Added `test_h5_installation()`.
+ 
+ * FUNCTIONALITY: Added `angular_distance()` and `list_distance_metrics()`.
+ 
+ * FIX: Added workaround for being able to read h5 files on Windows.
+
 # dynutils 1.0.2 (2019-03-21)
 
  * BUG FIX: Fix generation of error messages in `%all_in%`.